af_unix: rework maybe_add_creds() to allow sleep
As a preparation for the next patches we need to allow sleeping
in maybe_add_creds() and also return err. Currently, we can't do
that as maybe_add_creds() is being called under unix_state_lock().
There is no need for this, really. So let's move call sites of
this helper a bit and do necessary function signature changes.

Signed-off-by: Alexander Mikhalitsyn <aleksandr.mikhalitsyn@canonical.com>